Revert "use int for types.LongType"
This reverts commit aa2695cf2c6066cefcdf705f26cfc8e58f94c65f.

+++ stdlib/2.7/types.pyi
@@ -13,7 +13,7 @@ TypeType = type
 ObjectType = object
 
 IntType = int
-LongType = int
+LongType = long
 FloatType = float
 BooleanType = bool
 ComplexType = complex
